requirements.txt
requests~=2.25.1
Pillow~=8.1.0
instabot~=0.117.0
urllib3~=1.26.2
python-dotenv~=0.15.0